+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/PolyglotThreadInfo.java
@@ -40,9 +40,14 @@
  */
 package com.oracle.truffle.polyglot;
 
+import java.lang.management.ManagementFactory;
+import java.lang.management.ThreadMXBean;
 import java.lang.ref.Reference;
 import java.lang.ref.WeakReference;
 import java.util.LinkedList;
+import java.util.concurrent.TimeUnit;
+
+import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 
 final class PolyglotThreadInfo {
 
@@ -53,6 +58,10 @@ final class PolyglotThreadInfo {
     private int enteredCount;
     final LinkedList<Object> explicitContextStack = new LinkedList<>();
     volatile boolean cancelled;
+    private volatile long lastEntered;
+    private volatile long timeExecuted;
+
+    private static volatile ThreadMXBean threadBean;
 
     PolyglotThreadInfo(Thread thread) {
         this.thread = new WeakReference<>(thread);
@@ -66,9 +75,49 @@ final class PolyglotThreadInfo {
         return getThread() == Thread.currentThread();
     }
 
-    void enter() {
+    void enter(PolyglotEngineImpl engine) {
         assert Thread.currentThread() == getThread();
-        enteredCount++;
+        int count = ++enteredCount;
+        if (!engine.noThreadTimingNeeded.isValid() && count == 1) {
+            lastEntered = getTime();
+        }
+    }
+
+    void resetTiming() {
+        if (enteredCount > 0) {
+            lastEntered = getTime();
+        }
+        this.timeExecuted = 0;
+    }
+
+    long getTimeExecuted() {
+        long totalTime = timeExecuted;
+        long last = this.lastEntered;
+        if (last > 0) {
+            totalTime += getTime() - last;
+        }
+        return totalTime;
+    }
+
+    @TruffleBoundary
+    private long getTime() {
+        Thread t = getThread();
+        if (t == null) {
+            return timeExecuted;
+        }
+        ThreadMXBean bean = threadBean;
+        if (bean == null) {
+            /*
+             * getThreadMXBean is synchronized so better cache in a local volatile field to avoid
+             * contention.
+             */
+            threadBean = bean = ManagementFactory.getThreadMXBean();
+        }
+        long time = bean.getThreadCpuTime(t.getId());
+        if (time == -1) {
+            return TimeUnit.MILLISECONDS.convert(System.currentTimeMillis(), TimeUnit.NANOSECONDS);
+        }
+        return time;
     }
 
     boolean isPolyglotThread(PolyglotContextImpl c) {
@@ -78,9 +127,14 @@ final class PolyglotThreadInfo {
         return false;
     }
 
-    void leave() {
+    void leave(PolyglotEngineImpl engine) {
         assert Thread.currentThread() == getThread();
-        --enteredCount;
+        int count = --enteredCount;
+        if (!engine.noThreadTimingNeeded.isValid() && count == 0) {
+            long last = this.lastEntered;
+            this.lastEntered = 0;
+            this.timeExecuted += getTime() - last;
+        }
     }
 
     boolean isLastActive() {