+++ sdk/CHANGELOG.md
@@ -4,6 +4,8 @@ This changelog summarizes major changes between GraalVM SDK versions. The main f
 
 ## Version 19.3.0
 * The default temporary directory can be configured by [FileSystem](http://www.graalvm.org/sdk/javadoc/org/graalvm/polyglot/io/FileSystem.html#getTempDirectory--).
+* Added `org.graalvm.polyglot.ResourceLimits` that allows to specify context specific time and statement count execution limits.
+* Contexts can now be closed if they are still explicitly entered using `Context.enter` on the current thread. This allows for simpler error recovery code.
 
 ## Version 19.2.0
 * Added support for date, time, timezone and duration values in polyglot